Don't crash on invalid regex
When typing a regex it may be invalid until you are finished typing,
e.g. if you type [ it won't be valid until you finish it with ]. The
search kitten should not crash when this happens, but rather just remove
the mark, so you are able to continue typing.

There's still an issue with the error being printed (but it's cleared
right after, so it's not a huge problem), but I think we need a better
way to call the remote control to fix that.

Fixes #8