Fix compatibility with kitty v0.27.0
Call the kitty process to issue remote control commands instead of using
internal functions. This was noticably slower than the old method before
v0.27.0 which may have been the reason I went with using the internal
functions. However, in v0.27.0 it's fast (same speed as the old method
in v0.26.5 it seems).

Fixes #12